test(ssg): move the sidebar active-entry test to its own file
`rendering.rs` was already at the 350-line limit for changed sources.
