+++ crates/ox_content_ssg/src/html/tests.rs
@@ -27,6 +27,7 @@ mod image_gallery;
 mod lazy_widgets;
 mod mobile_css;
 mod mpa_navigation;
+mod nav_active;
 mod navigation_state;
 mod rendering;
 mod social;
